[JAX Runner] Refactor jax runner to delegate structured decoding logic  [5/N]
